Uploading in-progress work
This is the beginnings to a program that will be using forms of
correlation between motifs and continuous  bouts of song to compare
birdsongs to pre and post scenarios. However it is still very much in
the early days and is more currently serving as an exercise to
familiarise myself with Python and Github.