Experimentation with NeuralNetorks
Started using PyBrain, intend to apply it to syllable detect for me
using a simple feedforward neural network to recognise the spectrogram
of the song and pick out areas of syllables. This will result in a
running value that can be used as a "song characteristic" of what could
be viewed as "syllableness", this, with conjunction of amplitude and
wiener entropy thresholding may provide a robust mechanism for perfect
syllable extraction with little manual input.