MicSpeakerPro
Changed name of project, reminescent of SAP2011. Also Created new folder
to hold all the versions as they come in.
Plenty of progress, especially in automating. First half is done where u
give it a wav directory, and it will systematically play through all
files and record, leaving the output in another defined folder ready for
SAP to grab and create features from. The input of these features has
also been added in the program, and the alignment and correlation has
started. This final part is yet to be automated.